#29 fix subquery evaluation in FederationSail due to wrong join order
The evaluation of subqueries in FederationSail was not correct as
revealed by W3C test sq14. This is caused by an unsuitable join order,
where the sub select is processed using the results of a graph pattern
as intermediate result.

The simplest fix for this case is to set the cardinality hints for the
optimizer for SLICE algebra nodes to 1 such that they are executed as
left join argument with high probability.

The full fix (which will be done in another change) could to use a hash
based join strategy.