#29 fix suqbquery evaluation in FederationSail (use hash join)
In certain cases the nested loop join approach is not applicable (e.g.,
if the right argument is a sub-select with ORDER BY or LIMIT, see sq14).

This change now applies a HashJoin strategy in such cases.

Note that the HashJoinIteration has to be extended such that the left
intermediate result iteration can be passed to it.