issue #262: add a test and fix the issue in the HTTP server with empty prefixes
The issue was caused by the use of String.split(...) to decompose the
request path into its components (using the slash '/' as separator),
in order to extract the last component, which indicates the prefix the
request was about: in case of an empty prefix, the last component
would be an empty string, which is discarded by String.split(..)

The proposed solution is to use String.lastIndexOf to detect the
position of the last slash in the request path, which indicates
where the relevant prefix begins.

As the original code, the proposed modification accepts URLs that are
not strictly conforming to the specified request format: e.g. any
number of intermediate components are silently ignored.

Signed-off-by: Manuel Fiorelli <manuel.fiorelli@gmail.com>