issue #614 : Console Federate command broken by log level debug
The Console Federate command is broken if log level is set to debug due to the incorrect use of log.isDebugEnabled for control flow.

Debugging this involved updating Mockito to 2.2.0 and Logback to 1.1.7, but those changes don't seem to be related to the solution so they can be backed out if legal can't approve by the time of the next planned release.

Fixes #614

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>