issue #665 : Add tests for upper and lower case language tags
Existing test was based on a mixed upper and lower case language tag. These two new tests augment that to ensure that all uppercase and lower case language tags are parsed without issues.

However, this does not rely on them being normalized, it just ensures that the abstract Model is consistent and round-trips from Model->InputStream->Model to ensure it can be parsed without errors.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>