issue #684 : Ensure that Transaction.hasActiveOperations gets a clean snapshot
Transaction now relies on synchronisation on the futures queue (smoother than a list for append-only) to make the ExecutorService.submit and the futures.add calls atomic.

There is an additional Semaphore to ensure that within hasActiveOperations we get a clean snapshot of the tasks occurring inside of the ExecutorService by making task executions almost completely mutually exclusive with hasActiveOperations.

The remaining case, where a task is about to complete, and has released its Semaphore, but the ExecutorService has not set isDone on the Future, is fine for our purposes, as it only extends the life of the Transaction rather than cutting it short.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>