issue #684 : Remove awaitTermination
We are using a sequential single threaded ExecutorService. By calling shutdown while we have a lock preventing other threads from submitting tasks, we are able to guarantee that the close operation is the last task to be executed, and hence we do not need to call awaitTermination as for multi-threaded ExecutorService implementations.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>