issue #93 : Add non-transactional versions of Repositories methods
Some Repository implementations do not silently succeed if transactions are attempted to be committed on read-only endpoints when there were no changes queued up for the transaction.

Hence, it is necessary to add versions of all of the Repositories methods that do not attempt to begin, commit, or rollback transactions.

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>