Issues/#1017 Fix an issue with ParsedIRI and IRIs with host starting with a digit
Currently in the parseHost() method if the host starts with a digit then we consider that the host
is IPv4 address. However, there are no restrictions on using digits in host names so the parseHost() method
throws exceptions on perfetcly fine host names which start with a digit.
The suggested approach is: deltect a leading digit in the host name -> try to parse as IPv4
If the parsing fails -> parse the host as a regular host name (ireg-name according RFC-3987).
Once the host is parsed check if it is valid (it is considered invalid if its TLD is not made of letters only).
If the host is invalid we consider that indeed IPv4 address was intended and we throw.

Signed-off-by: Yasen Marinov <yasen.marinov@ontotext.com>