#71: make elasticsearch integration test work
We need to use lucene 6.6.1 explicitly as there is a version check
(assertion) in elasticsearch 5.6.9.

In addition, the elasticsearch compliance test needs to be run with the
Java security manager disabled. Also the "JarHell" class needs to be
overridden with no-ops in order to make the test work with the classpath
and all the transitive dependencies.


Signed-off-by: Andreas Schwarte <aschwarte10@gmail.com>