#1635: introduce FederationContext to keep the state of the federation
The FederationContext is introduced to keep the state of the Federation.
As of this change we added the reference to the FederationManager to it,
and inkect it in the FedX sail and FedXConnection.

This avoids static access to the manager.

Note that in follow up changes further static lookups to the manager
will be replaced.

Signed-off-by: Andreas Schwarte <aschwarte10@gmail.com>