#1635: properly support external federated service resolvers
As of this change FedX properly supports (external) federated service
resolvers that are optionally set upon initialization time. If not set,
FedX uses the default SparqlServiceResolver.

The federated service resolver can be set using the FedXFactory or on
the FedXRepository prior to initialization.

Note that the implementation checks first if a SERVICE clause references
a federation member, only if it doesn't the SERVICE request is
delegated.

Signed-off-by: Andreas Schwarte <aschwarte10@gmail.com>