#1718: fix invalid cookie warnings in HttpClient from SPARQLRepository
It turns out that the default configuration of the HttpClient has a
configuration for cookie parsing which does not follow todays standards,
thus yielding to warnings in the log for certain cookies.

According to the HttpClient documentation users are advised to
explicitly use RFC 6265 compliant policies.

This is done as of this change by adjusting the CookieSpec in the
default HttpClientBuilder used in the RDF4J SPARQLRepository.

Signed-off-by: Andreas Schwarte <aschwarte10@gmail.com>