Fix vectored evaluation using VALUES in RepositoryFederatedService
This change adjusts the optimized evaluation of SERVICE queries using
the VALUES clause.

In order to address performance and correctness issues, the VALUES
clause is now inserted in the beginning of the WHERE block (instead of
appending it to the generated query).

This change also does some cleanup to the RepositoryFederatedService
implementation (e.g. replacing deprecated code)

Change is covered by numerous integration unit tests