GH-1816: improve (external) SERVICE evaluation in FedX engine
This change improves and fixes the evaluation of (external) SERVICE
clauses in FedX.

Before this change there have been issues with blocking behavior and and
hanging queries, when SERVICE queries were evaluated as bound joins. As
this was not the default, it was not yet discovered. The reason was that
HTTP slots in the background tuple results were not processed in a
suitable order.

As of this change we slightly adopted the evaluation logics: now we
basically consume the background tuple result and keep it in memory for
upcoming operators, thus process it in order.

Also the Bound Join evaluation is now activated by default.

Moreover, a slight follow up of #1802 is added: the boundJoinBlockSize
is no longer hard-coded in the RepositoryFederatedService, but now uses
the value from the field.