GH-1814: support to configure inference on federation queries
As of this change it is possible to configure inference on federation
queries using Query#setIncludeInferred. By default the value is set to
'true', similar to the SailQuery.

The corresponding value is now passed to the TripleSource and is thus
considered for all sub-queries sent to the relevant endpoints.

Functionality for the query types is validated in a unit test.