GH-1820: fix connection handling in RepositoryFederatedService
This change introduces a mode in RepositoryFederatedService which allows
to always use a fresh RepositoryConnection for the actual query
invocation. In order to properly close the connection, the result
iteration is wrapped in a "CloseConnectionIteration"

The new mode is active by default, i.e. as of this change a fresh
connection is used for each SERVICE invocation.

For the re-use of the managed connection we now made sure to support
synchronized access, i.e. avoid to loose the references to created
connections.

The unit tests show that connections are properly closed. If it wasn't
the case, we would observe hanging tests for > 30s (i.e. until the
automatic timeout occurs)