GH-1879: revise the cache implementation for source selection in FedX
This change deprecates the existing Cache interface and replaces it with
a new SourceSelectionCache.

Particularly, the previous implementation has various problems:

* the cache used persistence on disk with our an expiration strategy. In
a read-only environment this works fine, but if data changes in the
sources over time wrong results may be returned.
* cache persistence caused unexpected results also in an RDF4J workbench
use-case where the federation member repositories were populated only
after the federation was initialized
* the cache also maintained information for {?s, ?p, ?o} which clearly
requires invalidation
* on the technical level the cache uses Java object serialization


In order to have a clean design we now propose a designated interface.

We also provide a new implementation for the source selection cache
which under the hood uses a Guava cache. The cache spec is configurable,
however, we define a reasonable default (maxSize=1000,
expireAfterWrite=6h).

Note that the new implementation now actually has logic to infer cached
information when possible.

The new cache is also covered in a set of thorough unit tests.