GH-1882: basic support for property path in FedX engine
This change adds basic support for property paths in the federation
engine.

Example patterns:

* ?subClass rdfs:subClassOf/rdfs:label ?label
* ?class rdfs:subClassOf+ ?class
* ?concept rdfs:label|skos:altLabel ?label

Note that some adjustments to queryalgebra PathIteration were required:

* as in FedX we generate SPARQL query strings, variables must not
contain the "-" character (which is invalid for SPARQL queries). This
character is replaced with _
* in the FedX engine we are not seeing "QueryBindingSet" instances, but
"MapBindingSet" instances. The addBinding method is made robust to
support both instance types
* in remote settings we even see ListBindingSet. In the check we make
sure to convert the BindingSet instance if required

Note also that the cost model estimation now only shows a warning if the
type of the node is unknown.