GH-1882: optimization for property paths with an exclusive endpoint
This change provides an optimization for property path expressions that
yield results at a single endpoint. Similarly to the concept of
ExclusiveStatement we now use a marker for them.

We introduce the interface ExclusiveTuplExpr which can now be used to
declare any exclusive TupleExpr. Note that the existing
ExclusiveStatement and ExclusiveGroup nodes have also become instances
of ExclusiveTupleExpr.

For rendering such nodes in sub-queries the ExclusiveTupleExprRenderer
interface is defined. Additionally, some rendering logic is introduced
in the Query utility classes.

Note that ExclusiveGroups are now generalized such that they can contain
any ExclusiveTupleExpr (and not only ExclusiveStatements).

Notes:

* the if/else branches in StatementGroupAndJoinOptimizer were shuffled
around such that the instanceOf check is working correctly
(ExclusiveGroup is now an instanceof ExclusiveTuplExpr)