GH-1538: clean SparqlFederationEvalStrategy
Deprecate the UNION based evaluation of bound joins, which was not
active for a long time, and instead integrate the VALUES based bound
join implementation directly into SparqlFederationEvalStrategy.

This change integrates the implementation of
SparqlFederationEvalStrategyWithValues directly into its super class and
thus we can remove this class.

The configuration is adjusted accordingly to use the
SparqlFederationEvalStrategy.