GH-1538: restructure the FedX algebra interface hierarchy
This change restructures the FedX algebra interface hierarchy to share
common methods of all nodes.

The most top-level interface for any FedX specific algebra node is now a
FedXTupleExpr.

In addition, we have the new interface VariableExpr which can be used to
determine the free variables (e.g. during cost estimation).

Note that the logging severity for unknown nodes is reduced to DEBUG as
the fallback values are sufficient for most cases.