GH-1904: fix ThredPool initialization in FedX
In FedX we use configurable threadpools for executing join and union sub
queries, as well as other remote requests. Also the pool size is meant
to be configurable via FedXConfig setters.

However, due to a bug in the construction of the ThreadPoolExecutor the
size is always set to a most 10 threads. This is because we have an
unbound queue and set the minimumCoreSize to 10. In such setting new
threads are only created if the core size of the pool is not reached.

We now explicitly set the core pool size to the configured number of
workers, and in addition configure the pool in such a way that idle
threads can get destroyed. The idle timeout is set to 60s.