GH-1538: do not require existence of base directory at FedX construction
At construction time of the FedX factory we actually do not require the
existence of the base directory.

Actually the base directory is only required to exist in rare federation
setups (e.g. with a managed NativeStore). This is because FedX currently
does not persist files to disk. Note that this was recently changed by
switching to an in-memory cache.

To make the federation factory work with different implementations of a
RepositoryManager and to not require the existence of potentially empty
directories we no longer throw an Exception.