GH-1970: fix logger name in SPARQLProtocolSession
The SparqlProtocolSession uses this.getClass() as a name for the logger.

When having an extension of the class, this leads to misleading log
output. Particularly the logging levels are also harder to control
(w.r.t namespacing and packaging).

The general recommendation for logger names should be to use the class
name, at least it should adhere to the rdf4j package structure.