GH-1990: fix FedX initialization order for RepositoryManager environment
Due to a wrong initialization order in an environment where the
federation is initialized through a RepositoryManager, references to the
evaluation strategy instance are not available in the Endpoint.

This in turn caused an NPE in the filter evaluation.

This change fixes the issue by first initializing the manager, and only
after that the federation members.