GH-2238: fix join order optimization for bound variables through BIND (#2239)
We now explicitly consider the binding names of the BIND clause (i.e. an
Extension) as variable for the join. In this way, statements where such
bindings occur obtain a lower cost.

Toy example for illustration:

SELECT ?project ?label WHERE {
 BIND(:SomeThing AS ?type)
 ?project rdfs:label ?label .
 ?project rdf:type ?type .
}

With this change the type BGP gets evaluated before the label BGP.
Change is covered with a unit test.