GH-2624: support pass through of single source federated queries
This change provides an optimization to the FedX federation engine to
allow pass-through of query results for single source queries, if a
TupleQueryResultHandler was supplied (i.e. if the query was executed
through TupleQuery#evaluate(handler).

The implementation is done through a helper tuple expression (to
propagate information into the federation evaluation strategy. Also we
use a number of marker interfaces.

The implementation is covered by unit tests whose assertions are very
strict and will ensure to find errors if other parts of the framework
change.

Note that if no pass-through is possible, the same logic as before is
applied.

Signed-off-by: Andreas Schwarte <aschwarte10@gmail.com>