GH-2840 Implement an RDFParser and RDFWriter for Newline delimited JSON-LD
Introduce NDJSONLDParser that extends the JSONLDParser. It collects each JSON-LD line from the input, puts all JSON objects in a list, flattening them if necessary and pass further to the JSON API as one single JSON object.

Introduce NDJSONLDWriter that groups triples by subject and context and writes an JSON-LD line for each group, by delegating to the JSONLDWriter.

Signed-off-by: desislava.hristova <desislava.hristova@ontotext.com>