GH-2904 spring-boot based read-only sparql endpoint
This introduces a minimal Spring boot based sparql endpoint that
has no functionality other than implementing the sparql (read) protocol.

There is a controller that does the response logic and
an application that can be extended for tests.

It should be in the tools directory however, the spring-boot
conflicts with the existing spring-mvc usage.

Also we use Jetty as tomcat tends to not like '{}|' in urls,
which browsers per HTML5 spec don't escape.

Signed-off-by: Jerven Bolleman <jerven.bolleman@sib.swiss>