GH-3483: refactor federation execution to use a per-query strategy
This change refactors the execution of federated queries to use a
per-query instance of the FederationEvaluationStrategy.

The instance is passed via the QueryInfo object, which is already
specific to a given query execution.

Note that as a next step a proper factory mechanism needs to be put in
place in FederationManager and FederationContext (i.e. this change is
only the preparation, and it in fact still uses the same instance of the
strategy).