GH-3698 Gracefully handle idle connection shutdown
Add a retry strategy for `408` responses from HTTP server. A server may
- instead of silently closing a connection - instead send a `408`
response to the next request after the idle timeout. This leads to
unexpected behavior on user side:

- A connection that appears to be open (`isOpen()`) actually errors out
when sending a request and is then closed
- Even when getting a new connection (`repository.getConnection()`) this
connection may be a re-used connection from the connection pool and
errors out when sending a query since the server actually closed the
idle connection and responds with a `408`.

The retry handler will retry the request with a refreshed connection after
receiving a `408`. It tries this exactly once. The `408` is hence
gracefully handled by default.