GH-3711 Retry pooled, shut down connections
When a request fails because the underlying connection was shut down on
the server due to inactivity only once a retry is attempted.

If the connection is not empty, the next connection from the pool will
be used for the retry. However, this connection can be older than the
idle threshold on the server too. In that case we need to clean up this
pooled connection as well and retry with another connection.

In the worst case the pool is full of connections that are already idled
out on the server. We then need to clean the pool and retry with a fresh
connection.