Add support for custom query analyzers in Lucene and ES
The indexing analyzer is configurable, however, this is only half the
problem because if we query with a different analyzer than what's
indexed, we will end up missing a number of results (particularly as it
pertains to stopwords and stemming). For the cases where you want
to configure the indexing analyzer, we should also be allowed to
configure the query analyzer.