GH-4290 ensure introduced implicit group node is not dropped
When a query has an aggregate in the projection but no explicit group
clause, an implicit group clause is introduced. The logic that takes
care of scoping projection aliases such that they are available for
evaluation of the ORDER BY clause needs to take this into account: if
there is an implicit group, that group needs to be used as the extension
argument, instead of the order clause argument (which does not
contain the implicit group node).