GH-4983: fine-tune bind join implementation in FedX engine
This change applies fine-tuning to the bind join implementation on
certain parameters.

Note that the original implementations and benchmarks were done when
there was no SPARQL 1.1 around, and FedX was using UNION constructs for
the bind join implementation. Since a few year back it is using VALUES
clauses.

Changes:

- default bind join size is set to 25 (also reflected in documentation,
some unit tests require adjustments)
- no longer send small hard-coded sub-queries of size 3 for the first
requests and instead rely on configured bind join size
- no longer send the first request in a non-bound way by default (Note:
kept as internal configuration parameter in the implementation to allow
modifications in extensions)
- support to influence the bind join size in customizations