GH-5030: fix FedXDataset support for queries with FROM clauses
FedX supports interpret the Dataset being defined on the query: if it is
of type FedXDataset, the federation is reduced to the members defined by
FedXDataset#getEndpoints().

This is a nice means to evaluate a query on a subset of the federation
and is generally working fine.

When the query contains a FROM clause, however, the externally passed
Dataset is wrapped into a FallbackDataset (which internally keeps the
FedXDataset as primary one.

This change now inspects the FallbackDataset and unwraps it to reduce
the members. Note that for execution still the "FallbackDataset" is
used.