GH-5033: fix pushing of limits for simple ASK queries in FedX
This change makes sure to push limits for simple ASK queries with a
single statement patterns into the query.

The optimization is the same as applied for simple SELECT queries with a
LIMIT.

Rational: if the limit is not pushed, the federation engine will first fetch all data for the statement pattern and only then locally check if there is at least one, i.e it will cause performance issues and memory pressure when there are many triples matching the statement pattern (for instance millions of persons).