Add link to Jelly to about.md
Added a link to the Jelly serialization format implementation for RDF4J: https://w3id.org/jelly/jelly-jvm

Jelly is a binary RDF format with a strong performance focus. The implementation for RDF4J can be either used as a Maven dependency, or as a JAR plugin that you drop into your classpath.

Jelly was also implemented for Apache Jena. An implementation for rdflib is planned.