GH-5227: fix binding assigner optimizer in FedX
The federation optimizer was missing to execute the binding assigner
(which injects external bindings into the statement pattern). The
consequence was potentially incorrect results (due to source source
selection with partial knowledge) as well as sub-optimal source
selection

Issue is covered with a unit test, which is failing in two places prior
to this change.