GH-5229: fix left bind join in FedX for single binding input
This change fixes a situation that can incorrectly cause empty results.
It happens when the input of the left argument is a single binding set
and for special source selection situations (e.g. the right argument is
marked as ExclusiveStatement while the endpoint does not provide data)

To avoid the issue we also use the regular left join logic also for a
single binding set input, which can handle the situation properly.

Issue is covered with a unit test.