GH-5310 We can optimize aggregate operators only during the execution phase
This is because even for constants in projections, they might not be
executed and therefore the value would be different in the final
bindingset (an empty bindingset or not).

Optimize the min/max case by not evaluating per row but per iteration.

Work arround many testcases now failing due to QueryRoot being a more
consitent head of the TupleExpr, don't take these into account in the
Explanation. Leading to different plans being marked as failed later.

Signed-off-by: Jerven Bolleman <jerven.bolleman@sib.swiss>